Move debug code to different location
